+++ src/behavior/collapse-expand.js
@@ -45,7 +45,7 @@ module.exports = {
     try {
       this.onChange(item, collapsed);
     } catch (e) {
-      console.warn('G6 自 3.0.4 版本支持直接从 item.getModel() 获取源数据(临时通知，将在之后版本清除)', e);
+      console.warn('G6 自 3.0.4 版本支持直接从 item.getModel() 获取源数据(临时通知，将在3.2.0版本中清除)', e);
     }
     this.graph.refreshLayout();
   }