+++ src/layout/radial/radialNonoverlapForce.js
@@ -51,12 +51,22 @@ class RadialNonoverlapForce {
      * the node size
      * @type  {number}
      */
-    this.nodeSize = params.nodeSize || 35;
+    this.nodeSizeFunc = params.nodeSizeFunc;
     /**
      * the strength of forces
      * @type  {number}
      */
     this.k = params.k || 5;
+    /**
+     * if each circle can be separated into subcircles to avoid overlappings
+     * @type  {number}
+     */
+    this.strictRadial = params.strictRadial;
+    /**
+     * the nodes data
+     * @type  {array}
+     */
+    this.nodes = params.nodes;
   }
   layout() {
     const self = this;
@@ -79,6 +89,7 @@ class RadialNonoverlapForce {
   getRepulsion() {
     const self = this;
     const positions = self.positions;
+    const nodes = self.nodes;
     const disp = self.disp;
     const k = self.k;
     const radii = self.radii;
@@ -94,7 +105,7 @@ class RadialNonoverlapForce {
         let vecLength = Math.sqrt(vecx * vecx + vecy * vecy);
         if (vecLength === 0) vecLength = 1;
         // these two nodes overlap
-        if (vecLength < self.nodeSize) {
+        if (vecLength < (self.nodeSizeFunc(nodes[i]) / 2 + self.nodeSizeFunc(nodes[j]) / 2)) {
           const common = k * k / (vecLength);
           disp[i].x += vecx / vecLength * common;
           disp[i].y += vecy / vecLength * common;
@@ -107,25 +118,28 @@ class RadialNonoverlapForce {
     const positions = self.positions;
     const disp = self.disp;
     const speed = self.speed;
+    const strictRadial = self.strictRadial;
     const f = self.focusID;
 
-    disp.forEach((di, i) => {
-      const vx = positions[i][0] - positions[f][0];
-      const vy = positions[i][1] - positions[f][1];
-      const vLength = Math.sqrt(vx * vx + vy * vy);
-      let vpx = vy / vLength;
-      let vpy = -vx / vLength;
-      const diLength = Math.sqrt(di.x * di.x + di.y * di.y);
-      let alpha = Math.acos((vpx * di.x + vpy * di.y) / diLength);
-      if (alpha > Math.PI / 2) {
-        alpha -= Math.PI / 2;
-        vpx *= -1;
-        vpy *= -1;
-      }
-      const tdispLength = Math.cos(alpha) * diLength;
-      di.x = vpx * tdispLength;
-      di.y = vpy * tdispLength;
-    });
+    if (strictRadial) {
+      disp.forEach((di, i) => {
+        const vx = positions[i][0] - positions[f][0];
+        const vy = positions[i][1] - positions[f][1];
+        const vLength = Math.sqrt(vx * vx + vy * vy);
+        let vpx = vy / vLength;
+        let vpy = -vx / vLength;
+        const diLength = Math.sqrt(di.x * di.x + di.y * di.y);
+        let alpha = Math.acos((vpx * di.x + vpy * di.y) / diLength);
+        if (alpha > Math.PI / 2) {
+          alpha -= Math.PI / 2;
+          vpx *= -1;
+          vpy *= -1;
+        }
+        const tdispLength = Math.cos(alpha) * diLength;
+        di.x = vpx * tdispLength;
+        di.y = vpy * tdispLength;
+      });
+    }
 
     // speed
     positions.forEach((n, i) => {
@@ -142,13 +156,15 @@ class RadialNonoverlapForce {
         const limitedDist = Math.min(self.maxDisplace * (speed / SPEED_DIVISOR), distLength);
         n[0] += disp[i].x / distLength * limitedDist;
         n[1] += disp[i].y / distLength * limitedDist;
-        let vx = n[0] - positions[f][0];
-        let vy = n[1] - positions[f][1];
-        const nfDis = Math.sqrt(vx * vx + vy * vy);
-        vx = vx / nfDis * radii[i];
-        vy = vy / nfDis * radii[i];
-        n[0] = positions[f][0] + vx;
-        n[1] = positions[f][1] + vy;
+        if (strictRadial) {
+          let vx = n[0] - positions[f][0];
+          let vy = n[1] - positions[f][1];
+          const nfDis = Math.sqrt(vx * vx + vy * vy);
+          vx = vx / nfDis * radii[i];
+          vy = vy / nfDis * radii[i];
+          n[0] = positions[f][0] + vx;
+          n[1] = positions[f][1] + vy;
+        }
       }
     });
   }