+++ src/layout/random.js
@@ -24,8 +24,14 @@ Layout.registerLayout('random', {
     const nodes = self.nodes;
     const layoutScale = 0.9;
     const center = self.center;
-    const width = self.width;
-    const height = self.height;
+    let width = self.width;
+    if (!width && typeof window !== 'undefined') {
+      width = window.innerWidth;
+    }
+    let height = self.height;
+    if (!height && typeof height !== 'undefined') {
+      height = window.innerHeight;
+    }
     nodes.forEach(node => {
       node.x = (Math.random() - 0.5) * layoutScale * width + center[0];
       node.y = (Math.random() - 0.5) * layoutScale * height + center[1];