+++ src/shape/edges/polyline.js
@@ -1,5 +1,5 @@
 const Shape = require('../shape');
-const Util = require('../../util/index');
+const Util = require('../../util');
 const PolylineUtil = require('./polyline-util');
 const Global = require('../../global');
 
@@ -47,12 +47,15 @@ Shape.registerEdge('polyline', {
     return keyShape;
   },
   getShapeStyle(cfg) {
-    const color = cfg.color || Global.defaultEdge.color;
-    const size = cfg.size || Global.defaultEdge.size;
     const customOptions = this.getCustomConfig(cfg) || {};
     const { style: defaultStyle } = this.options;
     const { style: customStyle } = customOptions;
-    const style = Util.deepMix({}, defaultStyle, customStyle, cfg.style);
+
+    const strokeStyle = {
+      stroke: cfg.color
+    };
+
+    const style = Util.deepMix({}, defaultStyle, customStyle, strokeStyle, cfg.style);
     cfg = this.getPathPoints(cfg);
     this.radius = style.radius;
     this.offset = style.offset;
@@ -74,9 +77,8 @@ Shape.registerEdge('polyline', {
     }
     const path = this.getPath(points, routeCfg);
     const attrs = Util.deepMix({}, Global.defaultEdge.style, style, {
-      stroke: color,
-      lineWidth: size
-    }, cfg.style, { path });
+      lineWidth: cfg.size
+    }, { path });
     return attrs;
   },
   getPath(points, routeCfg) {
@@ -132,7 +134,10 @@ Shape.registerEdge('polyline', {
         const newLabel = this.drawLabel(cfg, group);
         newLabel.set('className', labelClassName);
       } else {
-        const labelCfg = cfg.labelCfg || {};
+        const { labelCfg: defaultLabelCfg } = this.options;
+        const { labelCfg: customLabelCfg } = this.getCustomConfig(cfg) || {};
+
+        const labelCfg = Util.deepMix({}, defaultLabelCfg, customLabelCfg, cfg.labelCfg);
         const labelStyle = this.getLabelStyle(cfg, labelCfg, group);
         /**
          * fixme g中shape的rotate是角度累加的，不是label的rotate想要的角度