+++ src/shape/nodes/ellipse.js
@@ -1,5 +1,6 @@
 const Shape = require('../shape');
 const deepMix = require('@antv/util/lib/deep-mix');
+const Util = require('../../util');
 const Global = require('../../global');
 
 /**
@@ -59,21 +60,13 @@ Shape.registerNode('ellipse', {
   labelPosition: 'center',
   drawShape(cfg, group) {
     const customOptions = this.getCustomConfig(cfg) || {};
-    const { style: defaultStyle, icon: defaultIcon } = this.options;
-    const { style: customStyle, icon: customIcon } = customOptions;
-    const style = deepMix({}, defaultStyle, customStyle, cfg.style);
+    const { icon: defaultIcon } = this.options;
+    const { icon: customIcon } = customOptions;
+    const style = this.getShapeStyle(cfg);
     const icon = deepMix({}, defaultIcon, customIcon, cfg.icon);
-    const size = this.getSize(cfg);
-
-    const rx = size[0] / 2;
-    const ry = size[1] / 2;
 
     const keyShape = group.addShape('ellipse', {
-      attrs: {
-        ...style,
-        rx,
-        ry
-      }
+      attrs: style
     });
 
     const { width, height, show } = icon;
@@ -162,6 +155,31 @@ Shape.registerNode('ellipse', {
         className: 'ellipse-mark-bottom'
       });
     }
+  },
+  /**
+   * 获取节点的样式，供基于该节点自定义时使用
+   * @param {Object} cfg 节点数据模型
+   * @return {Object} 节点的样式
+   */
+  getShapeStyle(cfg) {
+    const customOptions = this.getCustomConfig(cfg) || {};
+    const { style: defaultStyle } = this.options;
+    const { style: customStyle } = customOptions;
+    const strokeStyle = {
+      stroke: cfg.color
+    };
+    // 如果设置了color，则覆盖默认的stroke属性
+    const style = deepMix({}, defaultStyle, customStyle, strokeStyle, cfg.style);
+    const size = this.getSize(cfg);
+    const rx = size[0] / 2;
+    const ry = size[1] / 2;
+    const styles = Util.mix({}, {
+      x: 0,
+      y: 0,
+      rx,
+      ry
+    }, style);
+    return styles;
   },
   update(cfg, item) {
     const customOptions = this.getCustomConfig(cfg) || {};